begin refactoring code in separate module files
