+++ index.html
@@ -6,6 +6,8 @@
     <link rel="stylesheet" href="./main.css">
   </head>
   <body>
-    <script src="./src/index.js"> </script>
+    <script src="./bundle.js"> </script>
+    <!-- <script src="./src/index.js"> </script> -->
+    <!-- <script src="./src/add_and_delete_lists.js"> </script> -->
   </body>
 </html>
